Livingwind playerdisplay (#46)
Works, I also like the Bottom-Up idea

* Moved each file up to src/main to make room for src/test.

Changed some build configs so we can dist the jar with the assets.

* One more small change, moved everything into main/java to match the
standard gradle structure.

* Added a test case that can be run through IntelliJ or through the cmd
line. Now we can test things.

close #41

* Resolved the merge conflict and fixed some space sizes.

* Needed a visual representation of the player.
Renders the player at it's current space and simplified the rendering
of each Space by using the Group class.

close #45

* Moved the player pieces to the bottom of the tile instead of the top.
Also, when a player moves out of the space, each other player in the
space repositions to accommodate the change.